+++ lib/eva/uc/ucapi.py
@@ -2331,6 +2331,8 @@ class UC_REST_API(eva.sysapi.SysHTTP_API_abstract,
                 return self.groups(k=k, p=rtp)
             elif kind == 'history':
                 return self.state_history(k=k, i=ii, **props)
+            elif kind == 'log':
+                return self.state_log(k=k, i='{}:{}'.format(rtp, ii), **props)
             elif kind == 'props':
                 return self.list_props(k=k, i=ii, **props)
             elif kind == 'config':